pyproject.toml
[project]
name = "sdlc-agents"
version = "0.1.0"
description = "Multi-agent AI system operationalising the SDLC through specialised Claude-based agents"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "anthropic>=0.40",
    "pydantic-ai>=0.0.13",
    "langgraph>=0.2",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    "mcp>=1.27.0",
    "python-frontmatter>=1.1.0",
    "anyio>=4.0",
]

[project.scripts]
sdlc-mcp-model = "src.tools.mcp_model_server:main"
sdlc-mcp-registry = "src.tools.mcp_registry_server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-bdd>=7.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-bdd>=7.3",
]
